Add files via upload
See NOTES_ON_THIS_VER.txt for important info

This version should work for mariadb/public_html/etc. but it's connected to my database. To change it for yours, it's a lil bit of effort bc you have to go thru a few of the files to change the database name u wanna use. I'll see if I can do anything abt this but rn I'm dead tired lmao